Working POC (or POS for what it's worth)
Now we have a working POC, it can run AOT code and its decently fast
however we still need to do some stuff like configuration files,
channels and that pesky keyboard thing.

In other words there is still a LOT of work to do and we still need to
make sure we are doing it correctly. Although seen how this run maybe a
daemon it's not needed right away.