Keyboard events are now sent correctly.
Implementation is still a mess. So some refactoring is needed. But we
got tabulation working for flutter focus elements. So know besides the
basic input things, we can also do keyboard based navigation in flutter.

Keyboard handling is so confusing in the flutter engine,
but its now kinda working semi correctly.